Revert "Dance: consolidate constants"
This reverts commit 3ed4e923eca7fdcd7d8637c29d9639897ceb4910.
