+++ apps/src/dance/ai/constants.ts
@@ -1,0 +1,6 @@
+export const DANCE_AI_SOUNDS = [
+  'ai-select-emoji',
+  'ai-deselect-emoji',
+  'ai-generate-no',
+  'ai-generate-yes',
+] as const;
